Final Product scene
Made a scene named final prodct. To be renamed to Game name when that is decided
made Gamecontroller to prefab
. 
Add anything you want to add to the final product of the game. 
Dont push anything that doesnt work, or incomplete

NO TEST ALLOWED in this scene
NOTHING INCOMPLETE ALLOWED